Be consistent about how to specify images for ebook readers.
Figures 1, 2, and 3 did not display properly on Kobo e-book readers.
The Mathematica figure looked fine.  The only real difference between
these was that the Mathematica figure specified a width, and the others
did not.  Thus, this commit makes Figures 1, 2, and 3 specify the same
width as the Mathematica diagram.

Same thing for the cover page.