Add an inline table of contents to the EPUB and MOBI versions.
According to various sources such as the Calibre FAQ, an inline table of
contents is important for the MOBI format.  Calibre can generate one
automatically but it puts at the very beginning or end and it makes more
sense to me to put it just after the Prologue, as done here.